+++ cmd/core/ports/notification.go
@@ -2,12 +2,15 @@ package portsservice
 
 type (
 	Dto struct {
-		Name  string `json:"name"`
-		Event string `json:"event"`
+		Id    string `json:"id" bson:"_id,omitempty"`
+		Name  string `json:"name"  bson:"name"`
+		Event string `json:"event" bson:"event"`
 	}
 
 	NotificationService interface {
 		SendNotify(dto Dto) error
 		CheckNotify(msg string) (string error)
+		SaveNotify(Dto) string
+		FindById(id string) *Dto
 	}
 )