+++ cmd/main/container/container.config.go
@@ -1,6 +1,7 @@
 package container
 
 import (
+	adaptermongodb "go-clean-api/cmd/infra/adapters/mongodb"
 	database "go-clean-api/cmd/infra/adapters/mysql"
 	amqpclient "go-clean-api/cmd/infra/integrations/amqp"
 	grpc_client "go-clean-api/cmd/infra/integrations/grpc"
@@ -8,15 +9,17 @@ import (
 	cache_client "go-clean-api/cmd/infra/repositories/cache"
 
 	"github.com/jinzhu/gorm"
+	"go.mongodb.org/mongo-driver/mongo"
 )
 
 type (
 	containerConfig struct {
-		Database     *gorm.DB
-		Grpc_client  grpc_client.GRPCClient
-		Amqp_client  amqpclient.AmqpClient
-		Http_client  http_service.HttpClient
-		Cache_client cache_client.CacheClient
+		Database      *gorm.DB
+		DatabaseNoSql *mongo.Database
+		Grpc_client   grpc_client.GRPCClient
+		Amqp_client   amqpclient.AmqpClient
+		Http_client   http_service.HttpClient
+		Cache_client  cache_client.CacheClient
 	}
 )
 
@@ -26,10 +29,11 @@ func newContainerConfig() containerConfig {
 	db.ConnectToDatabase()
 
 	return containerConfig{
-		Database:     db.Db,
-		Grpc_client:  grpc_client.New(),
-		Amqp_client:  amqpclient.New(),
-		Http_client:  http_service.New(),
-		Cache_client: cache_client.New(),
+		Database:      db.Db,
+		DatabaseNoSql: adaptermongodb.GetClient(),
+		Grpc_client:   grpc_client.New(),
+		Amqp_client:   amqpclient.New(),
+		Http_client:   http_service.New(),
+		Cache_client:  cache_client.New(),
 	}
 }